# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create a task under a project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TasksController_list",
        "summary": "List tasks (auto-scoped by role)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "project",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status name(s) from the project's workflow (custom or system default)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "P1",
                  "P2",
                  "P3"
                ]
              }
            }
          },
          {
            "name": "dueDateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueDateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "overdue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprintId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unassignedSprint",
            "required": false,
            "in": "query",
            "description": "true = only backlog tasks (no sprint assigned)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "issueType",
            "required": false,
            "in": "query",
            "description": "Filter by the project's issue type names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "An issue's parent (Epic-link or Sub-task's parent)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these labels",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "components",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these components",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "enum": [
                "dueDate",
                "priority",
                "status",
                "createdAt",
                "updatedAt",
                "rank"
              ],
              "type": "string"
            }
          },
          {
            "name": "customFieldFilters",
            "required": false,
            "in": "query",
            "description": "JSON-encoded array of {fieldId, value} - equals-match against a custom field, ANDed. Malformed/wrongly-shaped entries are silently dropped rather than rejected.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/overdue": {
      "get": {
        "operationId": "TasksController_overdue",
        "summary": "Paginated overdue task list (scoped by role)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "project",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status name(s) from the project's workflow (custom or system default)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "P1",
                  "P2",
                  "P3"
                ]
              }
            }
          },
          {
            "name": "dueDateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueDateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "overdue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprintId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unassignedSprint",
            "required": false,
            "in": "query",
            "description": "true = only backlog tasks (no sprint assigned)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "issueType",
            "required": false,
            "in": "query",
            "description": "Filter by the project's issue type names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "An issue's parent (Epic-link or Sub-task's parent)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these labels",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "components",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these components",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "enum": [
                "dueDate",
                "priority",
                "status",
                "createdAt",
                "updatedAt",
                "rank"
              ],
              "type": "string"
            }
          },
          {
            "name": "customFieldFilters",
            "required": false,
            "in": "query",
            "description": "JSON-encoded array of {fieldId, value} - equals-match against a custom field, ANDed. Malformed/wrongly-shaped entries are silently dropped rather than rejected.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/my-tasks": {
      "get": {
        "operationId": "TasksController_myTasks",
        "summary": "Tasks assigned to the current user",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "project",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status name(s) from the project's workflow (custom or system default)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "P1",
                  "P2",
                  "P3"
                ]
              }
            }
          },
          {
            "name": "dueDateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueDateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "overdue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprintId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unassignedSprint",
            "required": false,
            "in": "query",
            "description": "true = only backlog tasks (no sprint assigned)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "issueType",
            "required": false,
            "in": "query",
            "description": "Filter by the project's issue type names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "An issue's parent (Epic-link or Sub-task's parent)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these labels",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "components",
            "required": false,
            "in": "query",
            "description": "Filter to tasks with any of these components",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "enum": [
                "dueDate",
                "priority",
                "status",
                "createdAt",
                "updatedAt",
                "rank"
              ],
              "type": "string"
            }
          },
          {
            "name": "customFieldFilters",
            "required": false,
            "in": "query",
            "description": "JSON-encoded array of {fieldId, value} - equals-match against a custom field, ANDed. Malformed/wrongly-shaped entries are silently dropped rather than rejected.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/search": {
      "get": {
        "operationId": "TasksController_search",
        "summary": "JQL-lite compound search (Search/Dashboards v2)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "jql",
            "required": true,
            "in": "query",
            "description": "A JQL-lite compound query - see jql.util.ts for the supported grammar.",
            "schema": {
              "example": "project = \"665f...\" AND status != Done AND assignee = currentUser() ORDER BY priority",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/search/autocomplete-fields": {
      "get": {
        "operationId": "TasksController_autocompleteFields",
        "summary": "The Issue Navigator's JQL field/operator/keyword metadata",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/search/autocomplete-values": {
      "get": {
        "operationId": "TasksController_autocompleteValues",
        "summary": "Distinct, currently-in-use values for a JQL field (Module 4)",
        "parameters": [
          {
            "name": "field",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "issueType",
                "status",
                "labels",
                "components"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/bulk-move-sprint": {
      "patch": {
        "operationId": "TasksController_bulkMoveSprint",
        "summary": "Move multiple tasks into a sprint, or back to the backlog (BRD 6.2)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMoveSprintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/bulk-assign": {
      "patch": {
        "operationId": "TasksController_bulkAssign",
        "summary": "Reassign multiple tasks at once (BRD 6.2)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAssignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/bulk-relabel": {
      "patch": {
        "operationId": "TasksController_bulkRelabel",
        "summary": "Add labels to multiple tasks at once (BRD 6.2)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRelabelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/bulk-status": {
      "patch": {
        "operationId": "TasksController_bulkStatus",
        "summary": "Transition multiple tasks to the same status at once (Module 5)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Project & Task Management API",
    "description": "REST API for a Jira/Trello-style project and task management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BulkMoveSprintDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sprintId": {
            "type": "object",
            "nullable": true,
            "description": "null moves every task back to the backlog"
          }
        },
        "required": [
          "taskIds"
        ]
      },
      "BulkAssignDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assignee": {
            "type": "object",
            "nullable": true,
            "description": "null to unassign every task"
          }
        },
        "required": [
          "taskIds"
        ]
      },
      "BulkRelabelDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "labels": {
            "description": "Labels to add to every selected task",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "taskIds",
          "labels"
        ]
      },
      "BulkStatusDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "example": "Done"
          }
        },
        "required": [
          "taskIds",
          "status"
        ]
      },
...
    }
  }
}